telegram-bot/requirements.txt
aiogram==3.22.0
aiosqlite==0.20.0
python-dotenv==1.0.0
aiohttp==3.10.0
